+++ packages/react-three-fiber/package.json
@@ -1,0 +1,61 @@
+{
+  "name": "react-three-fiber",
+  "version": "6.0.0-alpha.3",
+  "description": "A React renderer for Three.js (web and react-native)",
+  "keywords": [
+    "react",
+    "renderer",
+    "fiber",
+    "three",
+    "threejs"
+  ],
+  "author": "Paul Henschel",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/drcmda/react-three-fiber/issues"
+  },
+  "homepage": "https://github.com/drcmda/react-three-fiber#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/drcmda/react-three-fiber.git"
+  },
+  "collective": {
+    "type": "opencollective",
+    "url": "https://opencollective.com/react-three-fiber"
+  },
+  "main": "dist/react-three-fiber.cjs.js",
+  "module": "dist/react-three-fiber.esm.js",
+  "types": "dist/react-three-fiber.cjs.d.ts",
+  "react-native": "dist/native.js",
+  "sideEffects": false,
+  "preconstruct": {
+    "entrypoints": [
+      "index.ts"
+    ]
+  },
+  "scripts": {
+    "prebuild": "cp ../../readme.md readme.md"
+  },
+  "dependencies": {
+    "@babel/runtime": "^7.13.9",
+    "@size-limit/preset-big-lib": "^4.10.0",
+    "react-merge-refs": "^1.1.0",
+    "react-reconciler": "0.26.1",
+    "react-use-measure": "^2.0.3",
+    "resize-observer-polyfill": "^1.5.1",
+    "scheduler": "0.20.1",
+    "use-asset": "^1.0.2",
+    "utility-types": "^3.10.0",
+    "zustand": "^3.3.3"
+  },
+  "peerDependencies": {
+    "react": ">=17.0",
+    "react-dom": ">=17.0",
+    "three": ">=0.126"
+  },
+  "peerDependenciesMeta": {
+    "react-dom": {
+      "optional": true
+    }
+  }
+}
