--- src/web/Canvas.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react'
-
-function Canvas() {
-  return null
-}
-
-export { Canvas }
