--- src/web/__tests__/webEvents.test.tsx
@@ -1,37 +1,0 @@
-// describe('web events ', () => {
-//   it('can handle onClick hook', () => {
-//     expect(true).toBe(false)
-//   })
-
-//   it('can handle onContextMenu hook', () => {
-//     expect(true).toBe(false)
-//   })
-
-//   it('can handle onDoubleClick hook', () => {
-//     expect(true).toBe(false)
-//   })
-
-//   it('can handle onWheel hook', () => {
-//     expect(true).toBe(false)
-//   })
-
-//   it('can handle onPointerDown hook', () => {
-//     expect(true).toBe(false)
-//   })
-
-//   it('can handle onPointerUp hook', () => {
-//     expect(true).toBe(false)
-//   })
-
-//   it('can handle onPointerLeave hook', () => {
-//     expect(true).toBe(false)
-//   })
-
-//   it('can handle onPointerMove hook', () => {
-//     expect(true).toBe(false)
-//   })
-
-//   it('can handle onLostPointerCapture hook', () => {
-//     expect(true).toBe(false)
-//   })
-// })
